refactor: remove unused events from ethereum-governance
