--- ethereum-governance/src/shared/events/gate_seal_events.ts
@@ -1,4 +1,0 @@
-export const GATE_SEAL_SEALED_EVENT =
-  'event Sealed (address gate_seal, address sealed_by, uint256 sealed_for, address sealable, uint256 sealed_at)'
-
-export const GATE_SEAL_FACTORY_GATE_SEAL_CREATED_EVENT = 'event GateSealCreated (address gate_seal)'
