--- ethereum-governance/src/shared/events/motion_created_events.ts
@@ -1,9 +1,0 @@
-export const MOTION_CREATED_EVENT = `
-    event MotionCreated(
-        uint256 indexed _motionId,
-        address _creator,
-        address indexed _evmScriptFactory,
-        bytes _evmScriptCallData,
-        bytes _evmScript
-    )
-`
