--- ethereum-governance/src/shared/events/motion_enacted_events.ts
@@ -1,5 +1,0 @@
-export const MOTION_ENACTED_EVENT = `
-    event MotionEnacted(
-        uint256 indexed _motionId
-    )
-`
