--- ethereum-governance/src/shared/events/vault_events.ts
@@ -1,7 +1,0 @@
-export const TRANSFER_SHARES_EVENT = `
-    event TransferShares(
-        address indexed from,
-        address indexed to,
-        uint256 sharesValue
-    )
-`
